If a user sends a sticker, the corresponding emoji is sent to IRC. (#10)
* If a user sends a sticker, the corresponding emoji is sent to IRC.

Probably the best way to send a sticker to IRC is to send a direct URL to the sticker so IRC users can open it up in a web browser.
However, it does not appear as though there is a way to do that.

From what I can guess based on what I've read, what Telegram does when it gets a sticker is it sends the file_id to its servers, and the sticker gets downloaded directly.  Therefore, its not possible to get a static URL to the sticker... unless we want to host the thing ourselves.

The next best thing we can do is send the corresponding sticker emoji to IRC.  They might not get a cool sticker, but they'll get something!

Also, added this as an option.  There may be emoji haters out there.

* Used === and !== instea of == and != per review.