Set NTBA_FIX_319=1 in env.example (closes #88)
This resolves a deprecation warning given by node-telegram-bot-api about
the "automatic enabling of cancellation of promises". Given context from
the issue linked in the upstream project, setting this environment
variable will ensure backwards-compatibility with future releases. If we
need this feature, then we need to enable it when we need it.

Ref:

    https://github.com/yagop/node-telegram-bot-api/issues/319#issuecomment-324963294