docs: Add config file glossary
Part of issue #118.

This commit adds a new page for a configuration file glossary. This page
is a highly-detailed overview of the config file and what each setting
does in the `env.example` file. It is intended for advanced users.

For now, this commit does not pull the duplicated parts out of
`installation.rst`, but I plan to deprecate that page in a future pull
request.