docs: Add new page to explain how to deploy Teleirc
This commit is part of issue #118.

A new page is added specifically to explain how to deploy Teleirc. It
consolidates the **Usage** section from `installation.rst` and all of
the `using-docker.rst` page. This commit intends to make deployment
options more clear and easy to understand.

Some methods were removed, like the ArchLinux package. It is many
updates behind on the bot, so I think it's better to recommend people to
use directly from source where possible.