docs: Fix all Sphinx warnings by changing hyperlink targets
This commit changes all of the hyperlink targets for different
communities using Teleirc. Instead of listing an entry like this:

    - My community (Telegram | IRC)

I switched to writing them this like this:

    - My community (@mycommunity | #my-community)

This resolves all the Sphinx warnings when building the docs. Sphinx
doesn't like it when multiple hyperlink targets share the same text.
This commit mostly scratches an annoying itch when working with the
docs.