docs: Add new FAQ section with common troubleshooting steps
Closes #85. Closes #119.

This commit adds a new _Frequently asked questions_ page to the docs. I
populated it with two questions specific to Telegram:

* How do I find a chat ID for a Telegram group?
* I reinstalled Teleirc after it was inactive for a while. But the bot
  doesn’t work. Why?

The first question addresses issue #119. Hopefully this better explains
a common challenge with the Telegram API: how to find a chat ID number.

The second question addresses issue #85 with a problem we found when
reactivating a really old bot, specifically the one used in the
`@fedora` supergroup. This edge case is mentioned in the FAQ, so I think
it's fair to close this issue too.