Add contributing guidelines (closes #77)
This commits adds contributing guidelines into the project. The
contributing guidelines includes information on how to set up a
development environment (closing #77), how to make a good pull request,
and some transparency into the maintainer response time on new issues or
pull requests.

Hopefully, this makes Teleirc easier to contribute to. It also gives us
a starting point for people who want to get involved with the project
for the first time as a developer.