Always write project name as TeleIRC (for branding)
This commit renames all uses of 'Teleirc' in our project to 'TeleIRC'.
This makes the name of our project match the cases of our new logo.
While this is a minor change, it better establishes the project identity
when we go to print materials for things like booth exhibits (e.g.
stickers, brochure hand-outs, etc.).