Reverse #134: Do not render IRC=>Telegram messages as Markdown
This commit is a partial reversal of the bold username feature
introduced in #134. This disables the Markdown rendering change for
messages sent from IRC to Telegram. Instead of only the username field
being rendered as Markdown, the entire message string is rendered. This
causes some messages to not send over the bridge if incomplete Markdown
is used and also mangles some URLs.

Eventually it would be nice to revisit this, but for now, this fixes a
critical bug at the expense of removing a feature on one half of the
bridge.