issue templates: Change label 'discussion' -> 'needs triage'
This commit is a nitpick, but I changed the `discussion` label to
`needs triage`. I think `needs triage` inspires action to decide if a
new issue is accepted for development or not. "Discussion" is broad and
open-ended, and anything can be discussed for about as long as you like.
So, I figure `needs triage` is more direct with the intent that a new
issue needs to be considered or reviewed before it is accepted to the
development backlog.

Signed-off-by: Justin W. Flory <git@jwf.io>