docs: Support Markdown file type (closes #208) (#215)
* dependencies: Sphinx 2.0.1 -> 2.2.1

Signed-off-by: Justin W. Flory <git@jwf.io>

* docs: Bump version string v1.2.2 -> v1.3.3

This should be documented somewhere as part of the release process. I
didn't realize our docs were publishing for older versions.

Signed-off-by: Justin W. Flory <git@jwf.io>

* docs: Support Markdown file type (closes #208)

This commit adds Markdown support to our documentation. Files that end
in a `.md` extension will be detected by Sphinx as documentation to
render in `docs/`. We can write in whatever format we wish.

Closes #208.

Signed-off-by: Justin W. Flory <git@jwf.io>