ci: Add job to build and test docs
This commit switches our Travis CI builds over to a matrix build system.
In addition to performing the standard unit test suite for NodeJS, it
will also kick up a VM to build and test our docs. This is a fairly
lightweight way to guarantee that our docs are building.

Also, you will notice this commit is failing! That is expected. The
`recommonmark` dependency was not added in #215, but it was added in
PR #217. Once that PR merges, this CI test will pass again.

Signed-off-by: Justin W. Flory <git@jwf.io>