:memo: docs: Add Message State Diagram (closes #221)
Closes #221.

This commit adds a new docs page to our site with a digitized version of
the Message State Diagram we white-boarded a couple weeks back. The goal
of this document is to bring all developers on the same page about how
Messages pass through TeleIRC. While this helps new developers, it also
helps us stay focused by keeping to the expected states and avoiding
deviation.

Signed-off-by: Justin W. Flory <git@jwf.io>