:alembic: ci: Use golangci-lint for Go CI tests (closes #256)
This commit adds the golangci-lint tool to our Go Travis builds. It took
a while to figure out, but this works and successfully lints our code.
It provides some helpful errors when things are not working as they
should be.

At time of making a pull request, the tests are indeed failing. However,
this is helpful feedback for us to iterate on in new PRs. :smile: I
propose merging this through even though the tests are failing so we can
begin using this in all future changes.

Closes #256.

Signed-off-by: Justin W. Flory <git@jwf.io>