:card_file_box: ci: Cache Python packages for faster builds
While working on the CI improvements for Go in #262, I discovered a way
to improve the overall speed and performance of the Python job that
builds and tests our documentation. By piping the requirements to `pip`,
we can leverage the caching in Travis CI so it doesn't have to install
the docs dependencies every time a new build is made.

Not sure exactly how fast it will speed things up, but thought it was
helpful to do anyways since it wasn't that hard.

Signed-off-by: Justin W. Flory <git@jwf.io>

ci: Create requirements.txt outside of git working tree

This suppresses a Sphinx warning about an untracked `requirements.txt`
in the working directory for the docs:

`/home/travis/build/RITlug/teleirc/docs/requirements.txt: WARNING: document isn't included in any toctree`

Signed-off-by: Justin W. Flory <git@jwf.io>