:package: dependencies: Run `yarn upgrade` before last Node release
This commit bumps the dependencies for the NodeJS version of TeleIRC
v1.x.x to fix CVE-2020-7598:

    https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2020-7598

This will be the final commit to `master` before cutting the v1.5.0
release. This will get us ready for The Big Merge™ sometime on Saturday,
when the `go-port` branch is merged to `master`.

I have not personally tested these changes, but I'd like to test them in
our TeleIRC developer chat before merging this commit.

Signed-off-by: Justin W. Flory <git@jwf.io>