:female_detective: codeclimate: Initial commit (closes #288)
This commit adds CodeClimate back to the project, along with some
Go-specific lints and checks. This uses standard tools like `gofmt`,
`golint`, and `govet`. I'm not sure exactly what the output of these
checks will be, so this commit is an experiment to see what the results
are and how useful they are to us as a project.

    https://docs.codeclimate.com/docs/gofmt
    https://docs.codeclimate.com/docs/golint
    https://docs.codeclimate.com/docs/govet

Closes #288.

Signed-off-by: Justin W. Flory <git@jwf.io>