:page_with_curl: docs: Officially license content CC BY-SA 4.0
This commit brings our documentation content under an explicit license,
instead of de facto being committed on the basis of the MIT License. A
Creative Commons Attribution-ShareAlike license is a better format for
written content and material. It is also a Free Culture license. I
figured this was the most open-ended license to use for this content.

Signed-off-by: Justin W. Flory <git@jwf.io>