:woman_judge: legal: Switch TeleIRC v2.0.0 license to GPL-3.0
This commit changes the license of RITlug/teleirc from the MIT License
to a GNU General Public License. Merging this commit to `master` is
conditional on the consent of all active team developers. This commit
message will explain the rationale for this change and why I think it is
important.

Share improvements back upstream
================================

The README always said to please share improvements back upstream, but
we are using the MIT License, perhaps the most permissive of all
software licenses. In respect to the volunteer developer time and
energy, I think the GPL is a better position of the TeleIRC project's
values as an open source project.

Copyleft > permissive
=====================

Most notably, we switch from a permissive license to a copyleft license.
If a permissive license says "this code is open source", a copyleft
license says "this code is open source and will stay open source". I
think the GPL is an effective license for TeleIRC because the basic
fundamental copyleft essence of keeping changes to derivative works
under the same license.

Signed-off-by: Justin W. Flory <git@jwf.io>