:skier: Initial commit, TeleIRC Quick Start Guide
This is a merge of our two different confusing sets of documentation,
which was first pointed out by @aNullValue in the dev chat. It deletes
the two old pages, and the new page is a manual merge of the two
documents into one, except updated for v2.0.0 (so far).

Eventually, it will include more concrete instructions on how to
actually install and use the bot. I will complete that change in a
future commit/PR.

Part of #297, but does not close it.

Signed-off-by: Justin W. Flory <git@jwf.io>