:memo: 1st Docs Refresh: Structural refactor, add Quick Start Guide (#323)
* :package: docs dependencies: Update to Py3.8, latest package versions

Periodic version bump, just to keep everything clean and fresh.

Signed-off-by: Justin W. Flory <git@jwf.io>

* :abc: misc: Comment clean-up and minor refactoring

* Updated copyright year
* Added current maintainers to author metadata
* Refactor comments / spacing on some sections for better readability

Signed-off-by: Justin W. Flory <git@jwf.io>

* :up: docs: Update version/release to v2.x.x releases

This is tedious to always remember. Need to better document this
somewhere in a release guide. Maybe something to create once we get
through the v2.0.0 release and run a retrospective.

Part of RITlug/teleirc#297.

Signed-off-by: Justin W. Flory <git@jwf.io>

* :speaking_head: docs: Add callout for v1.3.4 docs

Adds a note callout to top of index page to make sure the reader is
looking at the correct version of our docs for whatever release they are
using.

Part of RITlug/teleirc#297.

Signed-off-by: Justin W. Flory <git@jwf.io>

* :skier: Initial commit, TeleIRC Quick Start Guide

This is a merge of our two different confusing sets of documentation,
which was first pointed out by @aNullValue in the dev chat. It deletes
the two old pages, and the new page is a manual merge of the two
documents into one, except updated for v2.0.0 (so far).

Eventually, it will include more concrete instructions on how to
actually install and use the bot. I will complete that change in a
future commit/PR.

Part of #297, but does not close it.

Signed-off-by: Justin W. Flory <git@jwf.io>

* :boom: :arrow_right: user docs: Update more file paths

This brings more user documentation into the folder namespace. Just to
keep things tidy and organized…

Signed-off-by: Justin W. Flory <git@jwf.io>