docs: Add specific deployment steps (closes #297) (#325)
* :superhero: docs: Add specific deployment steps (closes #297)

These steps are a quick runthrough that basically explains how the
Ansible Role works. But this gives a quick and easy pathway for people
who are running TeleIRC to deploy it on a bare-metal Linux distribution
with `systemd`. We can try to support other use cases later, but I think
this is sufficient level of depth for release day.

Closes #297.

Signed-off-by: Justin W. Flory <git@jwf.io>

* Address @Tjzabel feedback in #325 on `go build` step

Signed-off-by: Justin W. Flory <git@jwf.io>