:new: :control_knobs: irc: Clean up config options (fixes #113)
This commit reworks the IRC-side configuration options to support new
parameters and configuration options:

* **IRC_BOT_NAME**: IRC bot nick. Already existed, but now better
  documentation alongside other confusingly-named variables.
* **IRC_BOT_REALNAME**: `REALNAME` field support. Allows freeform text.
  Added bot advertisement as a default option if left unset.
* **IRC_BOT_IDENT**: Seldom used, but occasionally useful ident field.
  Overridden if NickServ credentials provided.
* **IRC_NICKSERV_USER**: Explicitly set a NickServ username for a bot.
  Previously used the `IRC_BOT_NAME`, but those are not always identical
  in all situations.

This commit also reworks the `env.example` file to better organize all
IRC options, in addition to adding the new ones. I also expanded and
reworked the config file glossary in the docs.

This Pull Request does not implement tests, but I would like to ask to
follow up on this in a subsequent Pull Request, preferably during the
upcoming testing sprint.

Fixes #113. cc: @alcir

Signed-off-by: Justin W. Flory (he/him) <jflory@unicef.org>