Work in progress implementation. Currently segfaults.
This is a mostly complete implementation, but still needs to be
debugged. Uploading an image to the Telegram group chat causes a
segfault:

```
panic: runtime error: invalid memory address or nil pointer dereference
[signal SIGSEGV: segmentation violation code=0x1 addr=0x8 pc=0x6dc502]
```

This is likely due to incorrectly using pointers since it is a new
concept to me coming from Python.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>