Check if IRC messages come from the expected channel
In some setups, teleirc might be running behind an IRC bouncer
managing a bot account that is a member of multiple channels. In these
cases teleirc will receive messages for channels that are not bridged
to Telegram.

To avoid accidentally forwarding these messages to Telegram, this
commit explicitly checks that the channel from which a message is
received matches the channel that is configured in teleirc.

Note that girc code itself fetches the channel name using this array
index in `IsFromChannel`.