Fix crashing on empty /me
When teleirc receives CTCP ACTION, it assumes that it has one byte
after i.e. space in "\x01ACTION \x01". But user could send just
"\x01ACTION\x01" and this would crash the bot.

Working solution to this problem is to not add a space yourself but
rather take the one from the message. This way bot will handle correctly
both empty and filled action.